Move single parameter analysis to utils.analysis
